Packages 2: Return of the forgotten pushes (#10)
* Remove middleware.js

Most of this middleware is better existing in express-auth.js where it has access to the managers.

* Delete AuthManager.js

All the auth fucntions need are better existing in the express-auth.js where they have access to the managers

* Have & enforce required fields in group manager model

Required fields are "id" and "users". Enforcement is done in new(), and set(). Having these fields be part of the model removes the need for individual sets and gets.

* Rename TokenManager to SessionManager and enforce model requirements

Required fields are "id" and "userID", and "expires". Enforcement is done in new(), and set(). Having these fields be part of the model removes the need for individual sets and gets.

* Change UserManager model to have required fields.

Required fields are "id" and "password". Enforcement is done in add(), and set(). Having these fields be part of the model removes the need for individual sets and gets.

* Create module exports

Many were copied in here from Middleware and AuthManager.

* Add express to dependencies

Not really dependent on it, but it can't be used without so kinda?